Fix symlink being created inside $OLD_MOUNT_PATH
Fix return code and the symlink being created *inside* of the directory
$OLD_MOUNT_PATH, if this directory is not empty.

Note that solely empty directories inside of the $OLD_MOUNT_PATH
directory are removed by this shell script, e.g. files, symlinks and
non-empty directories are not!

While there should be no files, symlinks and non-empty directories
inside the $OLD_MOUNT_PATH directory, his is not fail-safe, plus the
script then creates the symlink *inside* of the directory
$OLD_MOUNT_PATH!

This simple fix at least returns an error in this case (which is not
evaluated anyway, AFAICS) and does not create another file-system object
inside the directory $OLD_MOUNT_PATH then.